+++ src/index.css
@@ -5,10 +5,6 @@
   font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
 }
 
-a {
-  text-decoration: none;
-}
-
 .heading {
   font-size: 55px;
   margin: 33px 0 0 33px;
