+++ src/index.html
@@ -13,8 +13,8 @@
     <div class="scoresDiv scores"></div>
     <h6 class="Add-head">Add your score</h6>
     <form action="">
-        <input type="text" placeholder="Your Name" class="Name">
-        <input type="text" placeholder="Your Score" class="Score">
+        <input type="text" placeholder="Your Name" class="Name" required>
+        <input type="text" placeholder="Your Score" class="Score" required>
         <button type="submit" class="submit">Submit</button>
     </form>
 </body>
