+++ webpack.config.js
@@ -3,7 +3,11 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
   mode: 'development',
-  entry: './src/index.js',
+  entry: {
+    index: './src/index.js',
+    post: './src/postScore.js',
+    get: './src/getScore.js'
+  },
   devtool: 'inline-source-map',
   devServer: {
     static: './dist',
